wrap puppet to ensure provisioning succeeds
Puppet will return a non-zero exit code if there's a problem with
the run, which includes the possibility that the chronos_job
provider isn't available on the first run (because Puppet also
manages the httparty gem).

To work around this, we call a provision script that always exits
0. This isn't entirely necessary, but is a nice-to-have so that a
simple `vagrant up` provisions the entire demo environment in one
shot.