pin puppetlabs/centos-7.0-64-puppet to 1.0.1
Prior to this commit, we simply used the latest
puppetlabs/centos-7.0-64-puppet Vagrant box. In v1.0.2, the box was
upgraded to include Puppet 4, which is incompatible with Vagrant 1.7.2.

Since Vagrant 1.7.3 and 1.7.4 seem to have problems of their own, this
commit pins the box we're using in this environment to v1.0.1, with
Puppet 3.x.